pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-slm-chart-analysis"
version = "3.0.0"
description = "Hybrid AI system for extracting structured data from chart images"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "That Le", email = "thatle@example.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
keywords = ["chart", "analysis", "ocr", "yolo", "ai", "computer-vision"]

dependencies = [
    # Core
    "pydantic>=2.0",
    "omegaconf>=2.3",
    "python-dotenv>=1.0",
    
    # Image Processing
    "pillow>=10.0",
    "opencv-python>=4.8",
    "numpy>=1.24",
    
    # AI/ML
    "torch>=2.0",
    "ultralytics>=8.0",
    "paddlepaddle>=2.5",
    "paddleocr>=2.7",
    
    # PDF/Document Processing
    "pymupdf>=1.23",
    "python-docx>=1.0",
    
    # Settings management (API config)
    "pydantic-settings>=2.0",
    
    # Utilities
    "tqdm>=4.66",
    "loguru>=0.7",
]

[project.optional-dependencies]
data-factory = [
    # Data Collection Tools
    "arxiv>=2.1.0",
    "requests>=2.31",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "google-search-results>=2.4",  # SerpAPI for Google Search
    "roboflow>=1.1",               # Roboflow datasets
    "matplotlib>=3.8",             # Generate synthetic charts
    "scikit-learn>=1.3",           # Data splitting
    "google-genai>=1.0.0",         # New Gemini API SDK for Chart QA
]

dev = [
    # Testing
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    
    # Linting & Formatting
    "ruff>=0.1",
    "black>=23.0",
    "mypy>=1.5",
    
    # Type stubs
    "types-Pillow",
    "types-tqdm",
]

slm = [
    # SLM Support
    "transformers>=4.36",
    "accelerate>=0.25",
    "bitsandbytes>=0.41",
    "peft>=0.7",
    # Gemini API for Stage 4
    "google-generativeai>=0.3.0",
]

tracking = [
    # Experiment Tracking (optional)
    "wandb>=0.16",
    "tensorboard>=2.15",
]

api = [
    # API Server
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    # Task queue (async job processing)
    "celery[redis]>=5.3",
    "redis>=5.0",
    # Async DB (SQLAlchemy for job persistence)
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",          # SQLite async driver (dev)
    "alembic>=1.13",
]

demo = [
    # Demo UI
    "streamlit>=1.29",
    "plotly>=5.18",
]

all = [
    "geo-slm-chart-analysis[dev,slm,api,demo,tracking]",
]

[project.scripts]
chart-analysis = "interface.cli:main"

[project.urls]
Homepage = "https://github.com/thatlq1812/chart_analysis_ai_v3"
Documentation = "https://github.com/thatlq1812/chart_analysis_ai_v3#readme"
Repository = "https://github.com/thatlq1812/chart_analysis_ai_v3"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.txt"]

# ============================================================================
# Black Configuration
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | data
    | models
    | outputs
)/
'''

# ============================================================================
# Ruff Configuration
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG001"]  # Unused function arguments in tests

[tool.ruff.lint.isort]
known-first-party = ["core_engine", "interface"]

# ============================================================================
# MyPy Configuration
# ============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "PIL.*",
    "ultralytics.*",
    "paddleocr.*",
    "fitz.*",
    "omegaconf.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external resources",
    "gpu: marks tests requiring GPU",
]

# ============================================================================
# Coverage Configuration
# ============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
